find_combinations and evaluate_combination are now @staticmethods on MetaAgents with a clean break
