+++ tests/examples/test_alliance_formation_model.py
@@ -24,7 +24,7 @@ def test_alliance_model_records_overlapping_memberships(monkeypatch):
         ]
 
     monkeypatch.setattr(
-        "mesa.examples.advanced.alliance_formation.model.find_combinations",
+        "mesa.meta_agents.MetaAgents.find_combinations",
         fake_find_combinations,
     )
