+++ tests/test_meta_agents.py
@@ -5,11 +5,7 @@ import pytest
 from mesa import Agent, Model
 from mesa.discrete_space.cell_agent import CellAgent
 from mesa.discrete_space.grid import OrthogonalMooreGrid
-from mesa.meta_agents import (
-    MetaAgents,
-    evaluate_combination,
-    find_combinations,
-)
+from mesa.meta_agents import MetaAgents
 from mesa.meta_agents.meta_agent import MetaAgent
 
 
@@ -128,9 +124,14 @@ def test_evaluate_combination(setup_agents):
     def evaluation_func(agent_set):
         return len(agent_set)
 
-    result = evaluate_combination(tuple(agents), model, evaluation_func)
+    result = MetaAgents.evaluate_combination(tuple(agents), model, evaluation_func)
     assert result is not None
     assert result[1] == len(agents)
+
+    instance_result = model.meta_agents.evaluate_combination(
+        tuple(agents), model, evaluation_func
+    )
+    assert instance_result == result
 
 
 def test_find_combinations(setup_agents):
@@ -143,7 +144,7 @@ def test_find_combinations(setup_agents):
     def filter_func(combinations):
         return [combo for combo in combinations if combo[1] > 2]
 
-    combinations = find_combinations(
+    combinations = model.meta_agents.find_combinations(
         model,
         set(agents),
         size=(2, 4),
@@ -162,7 +163,7 @@ def test_find_combinations_allows_zero_value(setup_agents):
     def evaluation_func(agent_group):
         return 0.0
 
-    combinations = find_combinations(
+    combinations = model.meta_agents.find_combinations(
         model,
         agents,
         size=2,
@@ -179,7 +180,7 @@ def test_find_combinations_inclusive_tuple_size_bounds(setup_agents):
     def evaluation_func(agent_group):
         return len(agent_group)
 
-    combinations = find_combinations(
+    combinations = MetaAgents.find_combinations(
         model,
         agents,
         size=(2, 2),
@@ -192,7 +193,9 @@ def test_find_combinations_inclusive_tuple_size_bounds(setup_agents):
 def test_find_combinations_without_evaluation_func(setup_agents):
     """No evaluation function yields no combinations."""
     model, _agents = setup_agents
-    result = find_combinations(model, model.agents, size=2, evaluation_func=None)
+    result = model.meta_agents.find_combinations(
+        model, model.agents, size=2, evaluation_func=None
+    )
     assert result == []
